+++ src/pretix/base/models/orders.py
@@ -3408,21 +3408,19 @@ class InvoiceAddress(models.Model):
         return d
 
     def describe_transmission(self):
-        from pretix.base.invoicing.transmission import TRANSMISSION_TYPES
+        from pretix.base.invoicing.transmission import transmission_types
         data = []
 
-        for t in TRANSMISSION_TYPES:
-            if t.identifier == self.transmission_type:
-                data.append((_("Transmission type"), t.public_name))
-                for k, f in t.invoice_address_form_fields.items():
-                    v = self.transmission_info.get(k)
-                    if v is True:
-                        v = _("Yes")
-                    elif v is False:
-                        v = _("No")
-                    if v:
-                        data.append((f.label, v))
-                break
+        t = transmission_types.get(self.transmission_type)
+        data.append((_("Transmission type"), t.public_name))
+        for k, f in t.invoice_address_form_fields.items():
+            v = self.transmission_info.get(k)
+            if v is True:
+                v = _("Yes")
+            elif v is False:
+                v = _("No")
+            if v:
+                data.append((f.label, v))
         return data
