+++ src/pretix/control/templates/pretixcontrol/event/mail.html
@@ -117,6 +117,9 @@
                     {% blocktrans asvar title_order_custom_mail %}Order custom mail{% endblocktrans %}
                     {% include "pretixcontrol/event/mail_settings_fragment.html" with pid="custom_mail" title=title_order_custom_mail items="mail_text_order_custom_mail" %}
 
+                    {% blocktrans asvar title_order_custom_mail %}Invoice{% endblocktrans %}
+                    {% include "pretixcontrol/event/mail_settings_fragment.html" with pid="order_invoice" title=title_order_custom_mail items="mail_subject_order_invoice,mail_text_order_invoice" %}
+
                     {% blocktrans asvar title_download_tickets_reminder %}Reminder to download tickets{% endblocktrans %}
                     {% include "pretixcontrol/event/mail_settings_fragment.html" with pid="ticket_reminder" title=title_download_tickets_reminder items="mail_days_download_reminder,mail_subject_download_reminder,mail_text_download_reminder,mail_send_download_reminder_attendee,mail_subject_download_reminder_attendee,mail_text_download_reminder_attendee,mail_sales_channel_download_reminder" exclude="mail_days_download_reminder,mail_send_download_reminder_attendee,mail_sales_channel_download_reminder" %}
